requirements.txt
pandas
pendulum
mysql-connector-python
minio